\name{bxpMake}
\alias{bxpMake}
\title{Creating container to draw the boxplot}
\description{This function is to create a "container" with the proper structure to draw a boxplot.}
\usage{bxpMake(n = 4)}

\arguments{
  \item{n}{number of classes to draw the boxplot. }
}
\details{Default value is n=4.}
\value{Returns a "container" with the proper structure to draw a boxplot with n (default=4).}

\author{IAZI}
