\name{colorsel}
\alias{colorsel}

\title{
Function which gets the colors from GUI selection
}
\description{
According to the number of breakpoints passed it allows to select colors while running the map creation
}
\usage{
colorsel(breakpoints)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{breakpoints}{the number of colors to be selected}
}
\details{
This is used internally inside the IAZIMAP function to choose number of colors acoording to the breakpoints
}
\value{
Selects the colors
}

\author{
IAZI
}
