\name{FName}
\alias{FName}
\title{Specify the save path}
\description{This function  points to the location where files or plots are saved. }
\usage{FName (FileName="", FPath=FilePath)}
\arguments{
  \item{FileName}{the name in which the files or plots to be saved.}
  \item{FPath}{the path to be saved.}
 }
\details{If we use '.FilePath' instead the path remain unchanged during the session.}
\value{Exports the object to the specified SQL table or as a CSV file.}
\author{IAZI}
\examples{}
