\name{Goodies-package}
\alias{Goodies-package}
\alias{Goodies}
\docType{package}
\title{IAZI Functions for model building}
\description{This package consists of IAZI special functions for boxplot, histogram, export, lm etc. }
\details{
\tabular{ll}{
Package: \tab Goodies\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Updates: \tab Added functions IIT.ModPass, IAZIMAP and IAZIInstallPkg  \cr
Date: \tab 2013-11-11\cr
License: \tab IAZI\cr
}
Functions:\cr
bowplot, box2class, box3class, boxclass, boxplotSQL, bxpMake, CrashWarning, DBExport, estExport, FixHiLoScale, HiFixLoScale, HiLoScale, hist.factor, hist.factorP, modExport, objects.summary, PHS.bxp, PHS.CheckFactorsAfterCut, PHS.DistributionLog, PHS.DistributionPosX, PHS.DistributionX, PHS.getcex, PHS.lmcook, PHS.lmhat, PHS.lmres, PHS.makeFactorMatrix, PHS.match, PHS.Mod5Pass, PHS.panel.cor, PHS.panel.hist, PHS.rlmw, PHS.ValidPass, REPMDate, resboxchar, resboxcont, resboxcontfix, resboxdummy, sigExport, SQLCompatibility, SQLExport, IAZIColor, FName, IIT.ModPass, IAZIMAP,breakpoint,IAZICol,ModPass,colrsel,map.Swiss and IAZIInstallPkg
}
\author{
IAZI
\cr Maintainers: \cr TVK <karthik@iazi.ch> \cr
RPO <Poduval@iazi.ch>
\cr PVN <Nair@iazi.ch>
}

