\name{IAZICol}
\alias{IAZICol}

\title{IAZICol function for choosing colors in map}
\description{
This function creates the manual colors for maps depending on the breaks and style of the breakpoints.
}
\usage{
IAZICol(brkpts=NULL, MapType=MapType)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{brkpts}{the number of breaks to the data.}
  \item{MapType}{the type of the map to be created-either "Gemeinde" or "Knaton".}
}
\details{
This function is used inside the IAZIMAP function.
}
\value{}

\author{
IAZI
}

