\name{IAZISummary}
\alias{IAZISummary}

\title{ Summary of a data frame}
\description{
Creates the summary of a data frame}
\usage{
IAZISummary(x,na.rm=TRUE)
}

\arguments{
  \item{x}{a data frame.}
  \item{na.rm}{removes the NAs in the data frame. Logical, TRue or False.}
}
\details{Majorly used for Exploratory analysis.}
\value{
Returns the summary statistics of the data frame.
}
\references{
}
\author{
IAZI}

\examples{IAZISummary(mtcars)}