\name{PHS.getcex}
\alias{PHS.getcex}
\title{Character size in a panel}
\description{This function returns the character size used in the panel.cor function depending on the number of variables in the cor analysis.}
\usage{PHS.getcex(n)}
\arguments{
  \item{n}{an integer.}
}
\details{}
\value{Returns the character size.}
\author{IAZI}
\examples{
sapply(1:20, PHS.getcex)
}




