\name{PHS.panel.cor}
\alias{PHS.panel.cor}
\title{Preparing the pannel for correlograms}
\description{This is the panel FUNCTIONS for pairs scatter plots with correlations on the upper panel with size proportional to the correlations. This function filters the cases with NA without causing an error.}
\usage{PHS.panel.cor(x, y, digits = 2, prefix = "", use = "pair", ...)}
\arguments{
  \item{x}{x co-ordinate.}
  \item{y}{y co-ordinate.}
  \item{digits}{for formatting the text. default is 2.}
  \item{prefix}{value to be prefixed to the text.}
  \item{use}{default is pair.}
  \item{\dots}{other graphical parameters.}
}
\details{}
\value{Prepare the panel to plot.}
\author{IAZI}
\examples{
}




