\name{PHS.panel.hist}
\alias{PHS.panel.hist}
\title{Preparing the panel for correlogram with histogram }
\description{This is the panel FUNCTIONS for pairs scatter plots with histogram on diagonal.}
\usage{PHS.panel.hist(x, ...)}
\arguments{
  \item{x}{x co-ordinate.}
  \item{\dots}{more arguments.}
}
\details{}
\value{Prepare the panel to plot the histogram while plotting correlogram.}
\author{IAZI}
\examples{
}




