\name{PHS.ValidPass}
\alias{PHS.ValidPass}
\title{Checking for the Validity of the parameters in a regression pass}
\description{Check whether the parameters and the combinations of methods chosen in the pass are valid.}
\usage{PHS.ValidPass(Pass, Step)}
\arguments{
  \item{Pass}{specify which pass.}
  \item{Step}{specify the step.}
}
\details{Used in PHS.Mod5Pass( ) to check the validity of parameters.}
\value{It shows message when the pass failed.}
\author{IAZI}
\examples{
}