\name{REPMDate}
\alias{REPMDate}
\title{Date function in IAZI standard format}
\description{This function prints the date & time in the format YYYYMMDD.hhmmss.}
\usage{REPMDate(...)}
\arguments{
  \item{\dots}{ arguments.}
}
\details{}
\value{This function returns the date & time in the format YYYYMMDD.hhmmss.}
\author{IAZI}
\examples{REPMDate()}




