\name{sigExport}
\alias{sigExport}
\title{Exports the model sigma values}
\description{This function  exports the object to an SQL table or a CSV file.}
\usage{sigExport(X, Table = "", File = "", ...)}
\arguments{
  \item{X}{data set to be exported.}
  \item{Table}{destination table name.}
  \item{File}{destination file name.}
  \item{\dots}{other arguments.}
}
\details{}
\value{Exports the object to the specified SQL table or as a CSV file.}
\author{IAZI}
\examples{}




