\name{stdev}
\alias{stdev}
\title{Standard Deviation}
\description{Replacing sd() function with stdev().}
\usage{stdev(x, ...)}
\arguments{
  \item{x}{data}
    \item{\dots}{other arguments.}
}
\details{Calculates the standard deviation.}
\value{Return standard deviation.}
\author{IAZI}
\examples{x<-1:10
stdev(x)
}
